@@ -13,14 +13,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.0.0",
-    "@angular/core": "~13.0.0",
+    "@angular/common": "~13.2.5",
+    "@angular/core": "~13.2.5",
     "@angular/fire": "^7.2.1",
-    "@angular/forms": "~13.0.0",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
-    "@angular/service-worker": "~13.0.0",
+    "@angular/forms": "~13.2.5",
+    "@angular/platform-browser": "~13.2.5",
+    "@angular/platform-browser-dynamic": "~13.2.5",
+    "@angular/router": "~13.2.5",
+    "@angular/service-worker": "~13.2.5",
     "@capacitor/android": "3.4.3",
     "@capacitor/app": "1.1.0",
     "@capacitor/core": "3.4.1",
@@ -32,20 +32,20 @@
     "@ionic/storage-angular": "^3.0.6",
     "firebase": "^9.6.7",
     "leaflet": "^1.7.1",
-    "rxjs": "~6.6.0",
+    "rxjs": "7.4",
     "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.1",
+    "@angular-devkit/build-angular": "~13.2.5",
     "@angular-eslint/builder": "~13.0.1",
     "@angular-eslint/eslint-plugin": "~13.0.1",
     "@angular-eslint/eslint-plugin-template": "~13.0.1",
     "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.0.1",
-    "@angular/compiler": "~13.0.0",
-    "@angular/compiler-cli": "~13.0.0",
-    "@angular/language-service": "~13.0.0",
+    "@angular/cli": "~13.2.5",
+    "@angular/compiler": "~13.2.5",
+    "@angular/compiler-cli": "~13.2.5",
+    "@angular/language-service": "~13.2.5",
     "@capacitor/cli": "3.4.1",
     "@ionic/angular-toolkit": "^5.0.0",
     "@types/jasmine": "~3.6.0",
