@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { tap } from 'rxjs/operators';
+import { tap, take } from 'rxjs/operators';
 import { AuthService } from '../services/auth.service';
 
 @Injectable({
@@ -12,6 +12,7 @@ export class IsLoginGuard implements CanActivate {
 
   canActivate(): Observable<boolean> {
     return this.authService.isLogin.pipe(
+      take(1),
       tap((isLogin) => !isLogin && this.router.navigate(['/auth'])),
       tap((isLogin) => console.log('IsLoginGuard', isLogin))
     );
