@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { map, tap } from 'rxjs/operators';
+import { map, tap, take } from 'rxjs/operators';
 import { AuthService } from '../services/auth.service';
 
 @Injectable({
@@ -12,6 +12,7 @@ export class IsLogoutGuard implements CanActivate {
 
   canActivate(): Observable<boolean> {
     return this.authService.isLogin.pipe(
+      take(1),
       tap((isLogin) => isLogin && this.router.navigate(['/'])),
       map((isLogin) => !isLogin),
       tap((isLogin) => console.log('IsLogoutGuard', isLogin))
