@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { AuthService } from '../services/auth.service';
 
@@ -6,11 +6,13 @@ import { AuthService } from '../services/auth.service';
   selector: 'app-ajustes',
   templateUrl: './ajustes.page.html',
   styleUrls: ['./ajustes.page.scss'],
+  changeDetection: ChangeDetectionStrategy.Default,
 })
 export class AjustesPage {
   constructor(private auth: AuthService, private router: Router) {}
 
   logout() {
     this.auth.logout();
+    this.router.navigate(['/auth/login']);
   }
 }
